+++ source/views/settings/screens/overview/developer.js
@@ -1,8 +1,10 @@
 // @flow
 import {Sentry, SentrySeverity} from 'react-native-sentry'
 import * as React from 'react'
+import {Alert} from 'react-native'
 import {Section, PushButtonCell} from '@frogpond/tableview'
 import {type NavigationScreenProp} from 'react-navigation'
+import {isDevMode} from '@frogpond/constants'
 
 type Props = {navigation: NavigationScreenProp<*>}
 
@@ -14,9 +16,23 @@ export class DeveloperSection extends React.Component<Props> {
 		Sentry.captureMessage('A Sentry Message', {
 			level: SentrySeverity.Info,
 		})
+		this.showSentryAlert()
 	}
 	sendSentryException = () => {
 		Sentry.captureException(new Error('Debug Exception'))
+		this.showSentryAlert()
+	}
+	showSentryAlert = () => {
+		if (isDevMode()) {
+			Alert.alert(
+				'Sentry button pressed',
+				'Nothing will appear in the dashboard during development.',
+			)
+		} else {
+			Sentry.setEventSentSuccessfully(event => {
+				Alert.alert(`Sent an event to Sentry: ${event.event_id}`)
+			})
+		}
 	}
 
 	render() {