+++ CHANGELOG.md
@@ -53,6 +53,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Migrated our error handling to Sentry.
 - Enabled inline requires in the metro bundler
 - Updated references to ASC to be CAAS in dictionary and hours
+- Show alert after dev sentry send message or exception 
 
 ### Fixed
 - Fixed an issue where Fastlane was reporting build failures despite having skipped the build (#3215)